fix deadlock on Windows
Summary:
the monitor talks to the server over a pipe, created by `Unix.pipe`.

on mac and linux, writing to pipes can be non-blocking by setting a flag. on Windows, writing to anonymous pipes is always blocking: it uses `WriteFile`, which [blocks if it can't write](https://docs.microsoft.com/en-us/windows/win32/ipc/anonymous-pipe-operations). (note: [named pipes can be nonblocking](https://docs.microsoft.com/en-us/windows/win32/ipc/named-pipe-type-read-and-wait-modes))

Lwt wraps file descriptors, by calling `Lwt_unix.of_unix_file_descr`, and sets the nonblocking flag properly for you -- except on Windows, where it can't, because anonymous pipes are always blocking. And if you tell it the fd is nonblocking -- as we do with `~blocking:false` -- when it's not actually nonblocking -- as it's not on Windows -- then Lwt does writes in the main thread, which can end up blocking the main thread!

this led to a deadlock between the monitor and server: if a lot of LSP commands happen quickly (e.g. typing triggers lots of completion commands), the monitor writes a lot of messages to the server and eventually fills up the buffer and the writes start to block. meanwhile, the server is trying to respond to these messages, but it doesn't use Lwt, so it's _always_ sending blocking writes. since it's blocked writing to the monitor, it can't read the messages from the monitor to unblock the monitor, and vice versa.

Changelog: [fix] Fix a bug that could cause the server to become unresponsive on Windows

Reviewed By: samwgoldman

Differential Revision:
D35446725 (https://github.com/facebook/flow/commit/95208ee769e01f193ffa49c99e7e8aabd38f0a11)

------------------------------------------------------------------------
(from ebf54cc96ceae1dd39337ad813f8e2a4d0e32668)

fbshipit-source-id: 0c468b381c84b0b53d9adda5944ab45bceff2b85