Error on annotations nested inside of destructuring
Summary:
Currently, we parse annotations inside of destructuring and then ignore them. This is quite confusing to the user - for example I have seem many examples of:
```
const [val: T, setVal] = useState(x);
```
This invalid behavior should be an error instead. The logic is simple: if we are a pattern and have a parent and also have an annotation, then error.

By making this an error, we can prevent new instances of this pattern, and help eliminate it. In the future, we may end up deciding to either support annotations inside destructuring, or not parse them at all - but either way we should eliminate these current invalid instances first.

Changelog: [errors] Error on annotations nested inside of destructuring - these were always invalid, we just ignored them before

Reviewed By: SamChou19815

Differential Revision: D35304565

fbshipit-source-id: f9600356767b0bb82c4ebcbc94c6f00a37a34fa7