[old-env] Do not attempt to update env entry if we already know the reassignment is illegal
Summary:
Changelog: [fix] Do not add further errors when we already errored on the illegal reassignment.

This diff tries to make the old env more consistent with the new environment. The new environment models this as `NonAssigningWrite` and does not attempt to change the type environment except for adding the illegal reassignment error.

Reviewed By: mvitousek

Differential Revision: D35473470

fbshipit-source-id: 227b99fb34f7b3e24a2cbe006447e81a46b7932e