[new-env] Record reads and writes of RHS of instanceof even if it is not a refinement
Summary:
We are crashing when we are reading a value refined by instanceof, but the RHS of instanceof is not an expression we recorded a value or refinement.

e.g.

```
if (foo instanceof 3) {foo} // 3 is not read anywhere
if (foo instanceof immutable.Map) {foo} // immutable.Map is not refined previously.
```

In this diff, I will detect the cases when the RHS of instanceof is untracked in the name_resolver, and then adds a dummy read and write so that the entry exist. statement.ml is also needed to record the type for that entry.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D35321421

fbshipit-source-id: 174b711b7e5aa4c09e1d64ec89e0e5d438dd120c