[speculation] Consider a tvar benign if its root is the same as the ignored tvar's root
Summary:
Changelog: [internal]

When we have overloaded functions that only differ by return type. e.g.

```
declare function overload(key: boolean): string;
declare function overload(key: boolean): number;
```

The speculation process will use the argument types to disambiguate, and does not care about whether we can or cannot disambiguate on the return type. This is implemented by ignoring the return value tvar even when it's unresolved after considering a case. This works pretty well with old env.

However, in the new-env, we tend to unify a lot. e.g. in `const a = overload(key)`, we will unify the tvar of `a` with the tvar of return value of call of `overload`. This alone is still not a problem, if we already know the type of `key` when we are type checking `overload(key)`, then the speculation can complete without even considering the return type. However, it breaks down in the following case:

```
L1 | () => { const a = overload(key) }
L2 | declare var key: boolean;
```

In the above case, the speculation cannot make progress at all in L1, but only after we type checked L2. At this time, `a` already accumulated some bound and become unresolved. Since it's also unified with the tvar of the return value, so it also participates in the speculation process, causing the speculation algorithm unable to make progress on either resolving return type and ignoring it, which results in speculation-ambiguous errors.

This diff fixes that by considering whether a tvar should be ignored by comparing its root against the root of ignored tvar. This way, we will overcome the limitation described above, by still considering the tvar of `a` to be the same as the tvar of the function call return, which should be ignored.

Reviewed By: panagosg7

Differential Revision: D35472159

fbshipit-source-id: 380e2fc99caaf388587d0fb11e187826bbe25ae6