Full support for optional chaining
Summary:
This diff implements the full behavior of optional chaining in Flow. The correctness of this work heavily relies on the refinement proposition and changeset behavior of the previous diff, but surprisingly doesn't have to interact with it much directly.

The main change in this diff is introducing the `optional_chain` method, which is called rather than `expression` whenever we inspect something that may be an optional chain in a position where optional chain refinements can affect a scope--this typically being the target of some refinement test.

The optional chain proceeds backwards through the expression until it encounters an expression that is not a potential optional chain (i.e. any expression that isn't a OptionalMember or OptionalCall). As it goes, it adds the appropriate `NonMaybe` refinements on each optional key of the chain (via `member_expression_refinement` for optional members and directly for optional calls). The refinement machinery in the previous diff builds a proposition and changesets for these keys as needed. Null and void tests, rather than negating the `Undefined/Void/Maybe` refinements, now just negate the entire proposition if appropriate, so that the then case of `if(x?.a?.b === undefined) {} {}` correctly sees a large disjunction with no changeset, while the else case contains a number of refinements. This extra negation permutes the results of our test cases a bit by turning e.g. `Maybe` into `Not (Not Maybe)`, but Flow uses classical logic anyways.

A minor but significant change here is that statement.ml now populates the environment from optional chains with the filtered version of the chain's type, without including possible nulls. This is because the introduction of changesets means that in the null case, the environment won't contain an entry for the chain in the first place.

With this change we now pass the optional chain test case. We actually do better than the old test case, because previously a test like `if (x?.()) { }` wouldn't have `x` be refined in the then case even if `x` was unhavocable, because we used a kind of bespoke havocing that was unconditional, whereas now if `x` is unhavocable it will remain `NonMaybe` after the call. In order to maintain the current test case behavior, I added some code to make a few variables havocable that weren't before.

Reviewed By: SamChou19815

Differential Revision: D35322158

fbshipit-source-id: 8e8f1de99f71ef5ef547c8602045327aa4cfa68a