handle exceptions when server dies
Summary:
We see exceptions like this sometimes:

```
Uncaught async exception: Lwt_stream.Closed
Raised at Lwt_stream.create_with_reference.push in file "src/core/lwt_stream.ml", line 169, characters 60-72
Called from FlowServerMonitorConnection.Make.write in file "src/monitor/connections/flowServerMonitorConnection.ml", line 100, characters 24-53
```

this kills the monitor, which (a) prevents it from potentially starting a new server, and (b) prevents `flow lsp` from starting a new monitor, resulting in "server is stopped" in VS Code.

at the same exact time as this exception, we see the server exit, such as ".flowconfig changed in an incompatible way. Exiting."

the server exiting is detected by the monitor in one Lwt thread that's doing a `waitpid` and then calls `ServerConnection.close_immediately`. another Lwt thread is processing client commands and trying to send them to the server. the latter throws because the connection is now closed.

since `ServerConnection.close_immediately` is only called when we already know the server is gone, we're obviously already handling the server being gone in that codepath, and so don't need to also do anything if we can't write the command to the server. so we just catch and ignore the exception.

Changelog: [fix] Fix a bug responsible for some LSP "server is stopped" errors

Reviewed By: samwgoldman

Differential Revision: D35465056

fbshipit-source-id: 4a844afb64f823f75af604ac5a6db9b06d17668a