[new-env] Add tests to assert that `x instanceof Foo.Bar` actually refines x to `Foo.Bar`
Summary:
In D35321421 (https://github.com/facebook/flow/commit/97a719d4e2c499b9a15bc4ec130fb2971189ebff), I fixed the problem that the RHS of instanceof are not properly tracked in the new-env and causing a crash. In that diff, I fixed the problem by always tracking the RHS of instanceof in the new env.

It is suggested that it can be simplified by not adding the refinement at all if the RHS is not tracked as a Val. However, this turns out to miss the instanceof refinements. This diff adds the test so that we won't accidentally miss the refinement later.
Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D35481312

fbshipit-source-id: 39132f2e9f8588cafabcfc11c576b494adaa57cf