Add constrained write use op to new env errors
Summary: This diff fixes most of the discrepancies in the `constrain_writes` test case, and in so doing makes the new env behave much more like old env + constrain_writes. It does so by adding the `ConstrainedAssignment` use_op to post-inference constrained write checks in the new env in exactly the same way it does in the old env.

Reviewed By: SamChou19815

Differential Revision: D35328310

fbshipit-source-id: 9af9e63696a16921b3d2b1b78d67b9c4912907c2