SSA ID caching
Summary:
This diff caches results from the new-env's type_of_state based on the ssa_id of the analyzed val. This allows us to re-use these types instead of recomputing a type every time a binding is read.

Notice that we cache _before_ refining-- this is intentional. The id given is the id of the target of the optional refinement.

The next diff will cache the results of Refinements entirely so that refinements shared between phi nodes do not need to be recomputed. This is especially helpful in switch statements.

Reviewed By: SamChou19815

Differential Revision: D35506933

fbshipit-source-id: be6eea172b4d42b192b66f0d8c753f03855d0caf