Fix forgotten havoc, part 1: havoc_heap_refinements
Summary:
When there's a branch in the control flow, we rely on some tricky
juggling of state in order to get the right environments (with the
right refinements, initialization states, etc.) at the right places.

A key workhorse of that juggling is Env.merge_env, which we invoke
where control-flow paths join up (at a number of callsites directly in
statement.ml, and several more indirectly through Env.in_refined_env)
in order to merge what happened in two different incoming paths to
produce an appropriate environment for the outgoing path.

The work that Env.merge_env does is all independent between different
bindings and refinements; it iterates through them.  Usually most of
them weren't touched, and are the same on all the paths; so as an
optimization, it relies on being passed a set of bindings and
refinement keys (a Changeset.t) which is supposed to identify all of
the binding-entries and refinements that may differ between the envs.

To make that work, everything in the Env module that mutates the
current environment needs to ensure that the current global changeset
has the appropriate keys.

We had a bug where Env.havoc_heap_refinements, which we use in
particular at function and method calls, wasn't adding the havoced
keys to the changeset.  As a result, when we later merge environments
at the end of some enclosing control-flow construct, we could forget
to propagate the invalidation to the surviving environment.

Fix that by adding the keys for any refis we find here to invalidate.

Along the way, simplify Scope.havoc_all_refis by dropping its unused
`?name` parameter.  There's never been a caller that actually passed
that parameter, and without it the function better matches its name
and becomes very simple, easy to reason about.  The simpler interface
also lets us write a clean, concise comment documenting what it's for.

Fixes: https://github.com/facebook/flow/issues/8251

Pull Request resolved: https://github.com/facebook/flow/pull/8781

Reviewed By: jbrown215

Differential Revision: D32400348

Pulled By: SamChou19815

fbshipit-source-id: e3ccdcf248eb5182af9787ea39ab5a50b3a22886