remove unused ocaml_overrides module
Summary: in theory we should be passing `-open Ocaml_overrides` so that every caller of `Sys` and `Unix` get shimmed with `Disk` so that we can inject test versions. but we don't, and we don't have any immediate plans to write more `Disk`-based tests, so let's just clean it up.

Reviewed By: samwgoldman

Differential Revision: D35498407

fbshipit-source-id: 8655a746aef5a082fde0b38f0e6926a2f69ca6ce