better surface uncaught parsing exceptions
Summary:
uncaught exceptions during `Parsing_service_js.do_parse` can cause `Parse_fail` results even when `~fail:false`. `~fail:false` otherwise returns parse errors in the `Parsed` response.

instead, i made a new variant for uncaught exceptions to make this explicit. I also log it with `FlowEventLogger`. where we previously `failwith`'d with no details, we now re-raise the exception so we can see the full error. and where we previously reported it to the user, we continue to do so but with the full stack so we can do something about it if it gets reported.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D35512364

fbshipit-source-id: a77e71a7b2dd795c7142797c33b888447d603137