[parser] improve unexpected token errors in types
Summary:
historically, the parser didn't tell you what it was expecting when it failed. as we try to recover, we typically keep trying and failing on the same token repeatedly. we relied on error deduplication to just end up with one error for that token, since they all had the same generic message.

nowadays, we tell you what we were expecting more often than not. this leads to a weird scenario where we have multiple errors for the same token, but some of them tell you what it was expecting and some don't.

to be more consistent, this diff adds what we were expecting to the two places missing it in `Type_parser`.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D35532571

fbshipit-source-id: 8ea82a1910c8ee189b4e7347e36e1d8dfc7761fe