[parser] improve error recovery in object type properties
Summary:
improves parsing when you're in the middle of typing an object type's property, e.g.

```
type T = {
  foo // here
  bar: string,
}
```

before, we expect a `:` but find `bar`, skipping `bar`; then expect a type but find `:`, skipping `:`; then expect a comma or semicolon and find `string`, skipping `string`; then expect another property key but find `,`, skipping `,`.

further, the locations are all out of whack. the "foo" property's overall range covers `foo` and `bar`, but contains a made-up `any` type whose loc is the `:`, which is outside the parent range!

now, if we don't find a `:` then we give up parsing the property without consuming anything else, and use the key's loc as the value (still an implicit `any` type) location so that it's within the parent range. when it goes to the next property, it finds `bar` as expected and parses it properly.

this is similar to D34885414 (https://github.com/facebook/flow/commit/aca5c1114f08521a5d641a9a5a514421d2f54708) which improved object *values*.

Changelog: [parser] Improve error recovery when in the middle of adding a property to an object type literal

Reviewed By: samwgoldman

Differential Revision: D35529326

fbshipit-source-id: 5c6fb05a636c40a9930380fe663923e621c8c94d