Fix projections of switch cases
Summary:
When visiting a switch case, we currently visit the discriminant multiple times to in order to construct the right refinement for each case. At the end, in order to re-populate the env values with the correct version of the discriminant for the pre-switch case, we restore the original environment and visit the discriminant again.

However, it's possible for information from the cases to still leak into the original view of the discriminant, if the cases introduce heap refinements (when the discriminant didn't already have a heap refinement). This arises in a case like

```
declare var x: { y: "foo" } | { y: "bar" }

switch(x.y) {
  case "foo": break
  case "bar": break
}
```
Here, we'll visit `x.y` to compare it to "foo" and "bar" for each case, and at each point we introduce a heap refinement for `x.y`. This, as an unwanted side effect, creates a "temporary" inhabitant of the location of `x` and `x.y` in the env values, containing the refined writes for those keys. Of course, ultimately we want to record that when we visit `x` in the discriminant, it refers to an unrefined version of `x`, and we get that when we re-visit the discriminant in the original environment at the end--we overwrite the temporary version of the variable from the switch cases.

However, since in the original environment `x.y` doesn't exist as a key, we don't have anything to overwrite it with! As a result, it survives the switch statement, and we end up thinking that `x.y` refers to a refined version of a projection at that point.

This translates in the new env to reads of the discriminant within the switch statement as being tvars that never receive a lower bound! The discriminant `x.y` is considered to be some refinements applied to a projection at *that same location*, so we have a tvar cycle with no lower bounds, and any read of `x.y` comes from that location. So, a modification of the above program to the following will have no errors.
```
declare var x: { y: "foo" } | { y: "bar" }

switch(x.y) {
  case "foo": (x.y: empty); break // no error!!
  case "bar": break
}
```

This diff fixes this by actively removing potential heap refinements from the env values when no actual refinement is found, rather than just treating this as a no-op.

Reviewed By: SamChou19815

Differential Revision: D35539484

fbshipit-source-id: 3a4a9a6871f4054bd1e9a983861a7164e9cfe6c4