Make catch params explicitly any, rather than tvar with no LBs
Summary:
Currently, catch parameters are open tvars. Typically, they don't receive any lower bounds (unless they are written to), which makes them effectively empty.

Empty isn't really a reasonable understanding for this type. Ideally, things with empty type should be uninhabited, which is clearly not the case for catch params. We've discussed making them mixed in the past, which would IMO be a great solution but with a huge rollout difficulty.

Instead, this diff makes catch parameters always `any`. This more accurately describes the behavior of a variable whose type cannot be tracked by Flow.

This change also makes progress on the new_env--there was previously a difference in behavior when writing to a catch parameter, where the constrained write check would see the provider for the catch parameter as an empty tvar and raise an error for any assignment. This makes the `constrained_writes` test case closer to finished, though not totally done.

Reviewed By: jbrown215

Differential Revision: D35475511

fbshipit-source-id: fbf9c05798abb785f04ec5801ecc6e719f3bfba0