[new-env] Do not record write_id in Env_api.Refinement for `This` and `Super` val
Summary:
In the added regression test, I found a very surprising behavior where the type of `this` points to a type of `this` in a previous function with refinement on `this`. After some investigation, it's found that we accidentally cache `this` and `super` nested inside refinements.

In the particular added test, we have two refinements on this in two different functions, which are translated into env vals of the following form

```
Refinement {refi_id=123, write_id=456, write_val=This}
Refinement {refi_id=321, write_id=456, write_val=This}
```

Note that the write id for `This` val is the same, because this is a hack to track `this` and `super` refinement in the name resolver without fully modeling them.

When we are reading the first value, the write id and the type of `this` is incorrectly added to the cache, which caused the second read to get the wrong base type of `this`.

This diff fixes the problem by making the `write_id` on refinement value optional, and set them to None when the underlying write is `This` or `Super`

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35531482

fbshipit-source-id: b8d7fa2b605b9502d0693fb07d9862eea3d2c0f8