Filter discriminant of switch cases in name_resolver and statement.ml
Summary:
This diff addresses a performance issue for very large switch statements in the new env.

In the old env, when we test a switch discriminant against a case, the type we're testing is considered to have been filtered to not include the previous, failed cases. Previously in the new env, we instead would always be testing the original discriminant. This means that, if we're e.g. testing a disjoint union with `n` elements against a switch case with a case for each element, we'll do `n^2` tests in the new env compared to `n(n+1)/2` in the old env. This is significant enough to show a serious performance degradation in pathological cases.

We fix this by introducing the convention that, in switches whose discriminant is a member expression, the refined version of the base object of the discriminant for each case is recorded in the new env at the location of each case. This can be computed by the name resolver, and then used as the target of the test during checking.

This change necessitates a change to the matching props system. Because we're not visiting the discriminant in a conditional context anymore, we might not view the discriminant's base value as a write location anymore--and in fact, I'm not sure why we definitely could previously in the first place, given that it's functionally a read. This means that looking up its location in the loc_env might not find anything. To address this, we change the matching props behavior to be a regular new_env read, looking up the output type at the location rather than the (possibly missing) input.

Reviewed By: SamChou19815

Differential Revision: D35515392

fbshipit-source-id: 1c7f09e2c6db12f430b9968a571a53b4795ff2f4