Rename and refactor set_env_entry
Summary:
This diff splits the two parts of `set_env_entry` into two different functions. Those two parts are:
1. Unify the given type with the loc-env's entry for it
2. Subtype that type against the providers.

I split this up because the subtyping code adds in a constrained-write check that requires both arguments to be fully resolved. Since we call set_env_entry in places where assignments are not happening, we are unnecessarily fully resolving variables and flowing them to mixed.

In the next diff in the stack, I change callers that do not perform an assignment to use `unify_write_entry` instead of `assign_env_value_entry`

Reviewed By: SamChou19815

Differential Revision: D35557326

fbshipit-source-id: dedd2f0495b01c168baf24a1c19764d6d1e3ecd4