Functions in fully resolved env
Summary:
This builds support for functions in the fully resolved environment. Functions need to be handled differently based on whether they're return-type-annotated or not--if they are unannotated, we need to check the function body, while if they are annotated, we just need to extract the function signature. We can reuse existing functions in statement.ml to do this.

Minor change: we pull `Func_sig_intf.kind` inside the `S` module type because ocaml had trouble determining that `Func_sig_intf.kind` and `Func_sig.kind` were the same type.

Reviewed By: panagosg7

Differential Revision: D33323988

fbshipit-source-id: 597853d1da017023b7db5bc2791210b2ca6b61f4