Ban object spreading of numbers and strings
Summary:
We currently allow you to spread a number or string in an object literal. I don't think this is useful or expected. In the case of a string, it's also currently inaccurate (currently we model the spread of a string as something with no properties, but actually `{..."hi"}` results in `{0: "h", 1: "i"}`).

We continue to allow you to spread booleans, as that is useful in the following conditional pattern:

```
const x = {...(cond && {a: number})};
```

There are a couple of instances of people doing something like

```
const s: string = ...
const o = {...(s && {msg: s})}
```
But since we still allow `boolean`s, this is easily convertible to
```
const s: string = ...
const o = {...(s !== '' && {msg: s})}
```

Additionally, this change also brings `Object.assign` and object spread closer together in type-checking behavior, which is desirable since they have the same runtime behavior.

Changelog: [errors] Ban object spreads of numbers and strings

Reviewed By: panagosg7

Differential Revision: D35489886

fbshipit-source-id: 8fafa1079f9be1f38d25ad523810565462cf1134