Use quick_compare for loc comparisons
Summary:
Comparing concrete and abstract locations during Ty-simplification raises exceptions.
A repro is shown in the included test. This diff fixes this by using `ALoc.quick_compare`
for `Ty.t` comparisons.

The tradeoff here is that we're giving up some completeness in the case where a concrete
location is compared against an abstract one and the two locations correspond to the
same source location. This case was rather rare, since the reported crash was not that
common. In this case the resulting type might include some duplicate members in unions,
etc, which is benign.

Changelog: [fix] fix crash when comparing abstract and concrete locations when simplifying normalized types

Reviewed By: mroch

Differential Revision: D35617367

fbshipit-source-id: 0e3a5a58f00cbc97a1964deee07fd3a96290c441