add merge-array-elements flag to annotate-declarations codemod
Summary:
Adds a flag to `flow codemod annotate-declarations-experimental` so that
in code like
```
let x = [];
x = [1];
if (0 < 1) { x = [""]; }
```
Before the output of the codemod would be
```
let x: Array<number> | Array<string> = [];
```
This annotation is known to cause [speculation-ambiguous] errors, including
in simple examples like the above.

After this change, when `--merge-array-elements` is passed, we get
```
let x: $ReadOnlyArray<string | number> = [];
```
which does not cause new Flow errors.

Changelog: [internal]

Reviewed By: pieterv

Differential Revision: D35653354

fbshipit-source-id: 53ccedb7ec9b9ec2627887f5068dd74a25a7a0e0