[scope] Consider type annotations of function parameters to be outside of function scope
Summary:
The old environment will visit the type annotations of functions before getting into the scope of function parameter and body. While the new environment considers them to be in the scope.

The inconsistency was found when we are investigating performance regression of new-env, and we found that the it will lead to an infinitely recursive type in the following case in new env:

```
function f(x: typeof x) {}
```

This diff fixes the scoping rule of the new env to match the old env. We achieve that by first collect all the type annotations in function parameters into a list, and then remove those type annotations in the original place. Then we visit the type annotations in a list in the outer scope first, before entering the inner scope and visit the function parameters with type annotation stripped.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D35478747

fbshipit-source-id: 263fd6cb9955fa274a1ebf2a6ea32c82f45a770f