Updates and opassign in fully resolved env
Summary:
This diff adds support for updates and opassign (other than And/Or/Nullish assign). We consume the `name_def` data structure that represents syntactically a ++/-- operation or a += / -= / etc assignment, and having already ensured that any variables that these expressions depend on have been resolved, we can just run the same logic as statement.ml for these cases.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D33385729

fbshipit-source-id: e4119a5c87910c809c8cf86f8bb47e9f5ee3e29f