Perform literal subtype check and matching props check even if the predicate is not used for refinements
Summary:
Changelog: [error] We will now check whether `literal` is a subtype of `expr` in `expr === literal`, regardless of whether `expr === literal` can be applied as a refinement

Sometimes we are creating some predicates only to be dropped later. For example in `if (true && (false || x === 1) {}`, the predicate `x === 1` is dropped when we are merging to make the Or predicate. In this case, the predicate becomes useless for the refinement.

The old and the new env agrees on the uselessness of the predicate, but differs on what to do with literal subtyping check. The old environment will only perform these checks on predicates that are not dropped and useful for the refinement, while the new env will add the checks to the context unconditionally.

In order to make the migration to the new-env smoother, we need to make two envs agree on the same behavior. I think the new environment behavior makes more sense, since `x === 3` when `x` can only be `1 | 2` or a string should be errored anyways.

However, this diff does not make the old env and new env fully agree yet. The old env can only fire matching props checks for the form `simple_name.prop === literal`, while on the new env can fire for arbitrarily object sub-expressions. This will be addressed in a future diff.

Reviewed By: panagosg7

Differential Revision: D35511986

fbshipit-source-id: d1615300ffb8d08bf6807b70054dd362619ea07a