update tests/generic_escape to constrained-writes
Summary:
Generic escape errors that were formerly raised on updates to global variables, e.g.
```
let x = 1;
class C<X> {
  p: X;
  m() { x = p; }
}
```
now are raised as "cannot-assign" errors, as we they correspond to trying to assign
a generic value `p` of type `X` to the variable `x` which has a concrete type `number`.

There is one category of errors that are not thrown anymore, that corresponds to cases
where the inferred type for the global variable is `any`:
```
let x = (1: any);
class C<X> {
  p: X;
  m() { x = p; }
}
```
Before constrained-writes, this would still be a [generic-escape] error.

In the test, there are 2 such errors that are not raised anymore.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D35688845

fbshipit-source-id: 1d46cc95c59808b7e8e4760a0ccafa0769bb8330