Imports in resolved env
Summary:
This work builds support for imports in the fully resolved environment. As usual, this is achieved by refactoring statement.ml to expose the appropriate functions, and then calling env_resolution on them.

In this case, to simplify the refactoring we do make a visible change, in that given a `import * as X` statement, we now have the reason for `X` point to `X` rather than the overall statement. This is consistent with how other import statements work, so I'd consider it an improvement.

Reviewed By: panagosg7

Differential Revision: D33392969

fbshipit-source-id: f23c5e92bce9048ee83347146b74961d1faf20d5