Continue checking files with indeterminate module types
Summary:
Before this diff, an indeterminate module type would cause a file to "fail"
parsing, meaning the file would be considered untyped and would not be checked.

This is too strict. We can emit an error and continue to check the file instead.
This makes a bunch of code simpler (don't need to handle `Error _` cases) and
will also provide a better user experience -- people can still use IDE commands
on the file.

We need to be careful about one thing: the file sig tracks imports which we use
to build the dependency graph. An `export ... from "foo"` statement which
results in an error should still record the dependency on `"foo"`. This diff
makes sure that all dependencies are recorded.

Reviewed By: mroch

Differential Revision: D35328749

fbshipit-source-id: d4d11988922fdb404a873223a40b8c94dcb08187