Drop redundant indeterminate module type errors from type checking pass
Summary:
If a file has an indeterminate module type, we find that during parsing when we
make a file sig. This is a "fatal" error so we don't even type check the file.

In the next diff, I'm going to make the file sig error non-fatal, at which point
these errors start showing up. That would normally be fine, since duplicate
errors are filtered out, but the locations are not always consistent, so we
would sometimes emit the same error at different locations.

To avoid that, I'm just removing the errors from here.

Reviewed By: mroch

Differential Revision: D35331388

fbshipit-source-id: 338292ce89e1cae845100dbfc46864c89728c845