add test for tricky dir naming
Summary:
`Module_js` currently has this:

```
  let path_if_exists =
    let path_exists ~file_options path =
      file_exists path && (not (Files.is_ignored file_options path)) && not (dir_exists path)
    in
    fun ~file_options resolution_acc path ->
      let path = resolve_symlinks path in
      let declaration_path = path ^ Files.flow_ext in
      if
        Files.is_flow_file ~options:file_options path
        && (path_exists ~file_options declaration_path || path_exists ~file_options path)
      then
        Some path
      else (
        record_path path resolution_acc;
        None
      )
```

I was trying to figure out what the `not (dir_exists path)` was adding.  `Files.is_flow_file` makes sure `path` is not a directory, but we don't call that on `${path}.flow`.

all of these checks are very redundant and we should straighten them out, so adding a test makes sure we don't mess it up.

Reviewed By: samwgoldman

Differential Revision: D35499743

fbshipit-source-id: 4eac6d0708cc9929d5fd1086aa92dabc996aa9b7