Numbers, booleans, enums, etc. shouldn't be subtypes of interfaces (e.g. `interface {}`)
Summary:
What should be a subtype of `interface {}`? I think most people might expect that only object types and instance types can be subtypes of interfaces. However, currently all primitive types are also subtypes of `interface {}`, including numbers, booleans, strings, and so on. I don't think this behaviour is useful or expected.

- Booleans have no properties or methods (other than those inherited from `Object`). For sure they should be excluded
- Numbers have no properties, or methods that are commonly used that would make sense to describe using an interface. If you really want to do this you can use `| number`
- Enums members are opaque and shouldn't be subtypes of the empty interface, and we discourage all use of the enum object itself as this prevents inlining

My initial thought was that we should also exclude strings and perhaps arrays, but these implement `@iterator` and since we use an interface to describe iterables, we need to keep this to not break a variety of things for now

We already have the built in type `$NotNullOrVoid` which can be used to cover everything that is not `null` or `void` if that is desired.

When changing from a denylist of type in the match (e.g. previously `MixedT`, `VoidT`, `NullT`) to an allowlist, OCaml helpfully pointed out that the cases for objects, instances, and functions (statics) were already covered in the `match` (so you do not see me adding those cases here).

Changelog: [errors] Numbers, booleans, and enums are no longer subtypes of the empty interface `interface {}`

Reviewed By: SamChou19815

Differential Revision: D35711007

fbshipit-source-id: 75f6cafd967ab7c12cefd1ebb6453768f3dfd739