Do not flow union of specific into general in `Env.merge_env`
Summary:
The literal subtyping check and matching props check in the old env are both slightly broken. The issue is that the unannotated general type is being  widened by incorrectly refined specific types.

For example, in the following case:

```
function foo(o: ?"A") {
    let c = o;
    switch (c) {
        case "A":
        case 'b': // errors in new-env, no error in old-env
    }
}
```

If you hover on c in line in IDE, you will see `?"A" | string`, where `string` is coming from `case 'b'`. This causes the literal subtyping check to be useless. Similar issue exists in the following matching prop case:

```
function foo(o: ?{foo: "A"}) {
    let c = o;
    switch (c?.foo) {
        case "A":
        case 'b': // errors in new-env, no error in old-env
    }
}
```

The subtyping check is unnecessary. It causes us to produce some duplicate errors (as you can see in the removed errors in tests), since the subtyping check should already be done inside the merged branches. Therefore, it should be removed.

Changelog: [errors] More previously incorrectly missing errors might appear in expressions of the form `name === 'literal'` and `name.prop === 'literal'` complaining that `literal` is incompatible with the type of `name` or `name.prop`.

Reviewed By: panagosg7

Differential Revision: D35751212

fbshipit-source-id: aaeb6ffb96e7c043ae499b438e9666b8aa998cdd