fix flaky libs_are_implicitly_included test
Summary:
this test uses a `sleep` so sometimes fails... even though 2 seconds should be enough time.

I think querying `flow ls` satisfies the same goal, even if it is less direct. the test is meant to verify that the server is watching libs, and it's possible that `flow ls` lists files that it's not watching, but currently, they both use `CommandUtils.file_options` which is the key piece here.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D35747817

fbshipit-source-id: 3fbdd74ae2ca5ca5428d4565c38309fea55dbbda