[new-env] Use refined value for initialization for identifiers in object destructuring patterns
Summary:
Right now in the new env, when we encounter an object destructuring patterns like `{a:{b:c}} = obj`, we have no understanding that `c` refers to `obj.a.b.c`, which is problematic for refinements as `obj.a.b.c` might be refined.

In the old env in [`destrucuring.ml`](https://github.com/facebook/flow/blob/309dcc932686fcf36fbcfae2fd3f8343f306bb99/src/typing/destructuring.ml#L110-L130), we build up an member expression as we go deeper into the object pattern, and use that to track refinements. This diff mirrors the same strategy, except that we only query for existing refinements when we reached the leaf of the pattern, and then use that refined value to initialize the id. Before querying the refinement, we will read the synthetic object, so that all the refinements, including those on `object` subexpressions, will be recorded and picked up by `destructuring.ml`

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35268179

fbshipit-source-id: 737d216b1750710022d5acb70fae0f0f203d6569