skip subtyping check on function reassignment in old-env
Summary:
In the old environment the code
```
function f(): void {}
var f: number = 3;
```
produced two errors: a [name-already-bound] error, and an [incompatible-type] error
due to number being incompatible with a function type. The latter of these errors
is redundant.

The error was raised through unification that happens in `unify_declared_type`,
which is intended to allow forward references of declared types. This function is also
used for variable declarations; so if a variable declaration attempted to declare the name
of an existing function or class, `unify_declared_type` would be called trying to unify
the variable's type with the function or class type.

This diff skips the unification in the above case. The `loc <> value_assign_loc`
check makes sure that the initialization of a forward-declared type with its type
body actually happens, and only subsequent assignments are skipped.

Note that this change is only relevant to old-env, and its goal is to make it more
compatible with the new-env, which doesn't produce the redundant error.

One minor difference in behavior is that in
```
declare function x(): void;
const x = 1;
      ^
```
which is an [name-already-bound] error on the second line, we now infer `empty` as
the type at the position of `^`. Before we would infer the type `() => void`, which
seems confusing.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D35569634

fbshipit-source-id: 5a37420c9e70facf561bc9fccfadd599dc491a3b